Refactor the event listener on style Moving it on building.js, sharing the meshes variable Being able to aply the function to diff layers Still issue of buildings disappearing
