+++ src/main.js
@@ -1,6 +1,6 @@
 // https://github.com/iTowns/itowns/blob/master/examples/source_stream_wfs_3d.html
 
-import { update, buildingLayer } from "./models/building";
+import { update, buildingLayer, applyStyle } from "./models/building";
 import { picking } from "./models/connectDataToBuidlings"
 import { addOrthoLayer } from "./models/ortho";
 import { addElevationLayer } from "./models/elevation";
@@ -118,83 +118,22 @@ for (const layer of view.getLayers()) {
 }
 
 
-let meshes = [];
-// Je crois que je suis obligé de faire l'event listener ici pour avoir accès à la constante view
 // Fonction qui applique un style à la couche BD Topo
 document.getElementById("bouton_ok_style").addEventListener('click', () => {
-    meshes = [];
-    // On récupère les infos de style
-    const champ = document.getElementById("in_ch1Topo").value;
-    const valeur = document.getElementById("in_v1Topo").value;
-    const couleur = document.getElementById("in_co1Topo").value;
-
-    // Nouvelle fonction colorBuildings qui correspond aux infos
-    function colorBuildings(properties) {
-        let color = new itowns.THREE.Color();
-        if (properties[champ] === valeur) {
-            return color.set(couleur);
-        }
-
-        return color.set(0x555555);
-    }
-
-    // Pas changée Placing the data on the ground
-    function altitudeBuildings(properties) {
-        return properties.altitude_minimale_sol;
-    }
-
-    // Pas changée Extruding the data 
-    function extrudeBuildings(properties) {
-        return properties.hauteur;
-    }
-
-    // Pas changée 
-    function acceptFeature(properties) {
-        return !!properties.hauteur;
-    }
-
-    // Comme iTowns ne possède pas de modification dynamique de style, on supprime la couche et on la recrée
-    const layer = view.getLayerById('WFS Building');
-    view.removeLayer('WFS Building');
-    layer.delete();
-
-    // Source
-    const geometrySource = new itowns.WFSSource({
-        url: 'https://wxs.ign.fr/topographie/geoportail/wfs?',
-        version: '2.0.0',
-        typeName: 'BDTOPO_V3:batiment',
-        crs: 'EPSG:4326',
-        ipr: 'IGN',
-        format: 'application/json',
-        extent: {
+    applyStyle(
+        view,
+        'WFS Building',
+        'https://wxs.ign.fr/topographie/geoportail/wfs?',
+        '2.0.0',
+        'BDTOPO_V3:batiment',
+        'EPSG:4326',
+        'IGN',
+        'application/json',
+        {
             west: 0.67289,
             east: 0.74665,
             south: 45.17272,
             north: 45.2135,
         }
-    });
-
-    // Geometry Layer
-    const geomLayer = new itowns.FeatureGeometryLayer('WFS Building', {
-        batchId: function (property, featureId) { return featureId; },
-        onMeshCreated: function scaleZ(mesh) {
-            mesh.children.forEach(c => {
-                c.scale.z = 0.01;
-                meshes.push(c);
-            })
-        },
-        filter: acceptFeature,
-        source: geometrySource,
-        zoom: { min: 14 },
-
-        style: new itowns.Style({
-            fill: {
-                color: colorBuildings,
-                base_altitude: altitudeBuildings,
-                extrusion_height: extrudeBuildings,
-            }
-        })
-    });
-
-    view.addLayer(geomLayer);
+    );
 });
