+++ src/models/building.js
@@ -128,3 +128,76 @@ function acceptFeature(properties) {
 */
 
 
+// Menu de style
+export function applyStyle(view, id, serverURL, version, nameType, crs, ipr, format, extent) {
+    // On récupère les infos de style
+    const champ = document.getElementById("in_ch1Topo").value;
+    const valeur = document.getElementById("in_v1Topo").value;
+    const couleur = document.getElementById("in_co1Topo").value;
+
+    // Nouvelle fonction colorBuildings qui correspond aux infos
+    function colorBuildings(properties) {
+        let color = new itowns.THREE.Color();
+        if (properties[champ] === valeur) {
+            return color.set(couleur);
+        }
+
+        return color.set(0x555555);
+    }
+
+    // Pas changée Placing the data on the ground
+    function altitudeBuildings(properties) {
+        return properties.altitude_minimale_sol;
+    }
+
+    // Pas changée Extruding the data 
+    function extrudeBuildings(properties) {
+        return properties.hauteur;
+    }
+
+    // Pas changée 
+    function acceptFeature(properties) {
+        return !!properties.hauteur;
+    }
+
+    // Comme iTowns ne possède pas de modification dynamique de style, on supprime la couche et on la recrée
+    const layer = view.getLayerById(id);
+    view.removeLayer(id);
+    meshes = [];
+    layer.delete();
+
+    // Source
+    const geometrySource = new itowns.WFSSource({
+        url: serverURL,
+        version: version,
+        typeName: nameType,
+        crs: crs,
+        ipr: ipr,
+        format: format,
+        extent: extent
+    });
+
+    // Geometry Layer
+    const geomLayer = new itowns.FeatureGeometryLayer(id, {
+        batchId: function (property, featureId) { return featureId; },
+        onMeshCreated: function scaleZ(mesh) {
+            mesh.children.forEach(c => {
+                c.scale.z = 0.01;
+                meshes.push(c);
+            })
+        },
+        filter: acceptFeature,
+        source: geometrySource,
+        zoom: { min: 14 },
+
+        style: new itowns.Style({
+            fill: {
+                color: colorBuildings,
+                base_altitude: altitudeBuildings,
+                extrusion_height: extrudeBuildings,
+            }
+        })
+    });
+
+    view.addLayer(geomLayer);
+}
